Reorganize code to remove references across directories wherever possible.
