+++ sdk/burdock/data/local_csv_adapter.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from burdock.metadata.collection import CollectionMetadata
-from burdock.sql import PandasReader
+from burdock.reader.sql import PandasReader
 
 from .dataset_adapter import DatasetAdapter
