+++ sdk/burdock/metadata/__init__.py
@@ -1,0 +1,4 @@
+from .collection import CollectionMetadata
+
+
+__all__ = ["CollectionMetadata"]
