+++ sdk/burdock/sql/__init__.py
@@ -1,7 +1,6 @@
-from burdock.metadata.collection import CollectionMetadata
+
 from .private_rewriter import Rewriter
 from .private_reader import PrivateReader
-from burdock.reader.sql.pandas import PandasReader
 from .parse import QueryParser
 
 
@@ -10,5 +9,5 @@ def execute_private_query(reader, schema, budget, query):
     query = reader._sanitize_query(query) if hasattr(reader ,"_sanitize_query") else query
     return PrivateReader(reader, schema, budget).execute(query)
 
-__all__ = ["CollectionMetadata", "Rewriter", "QueryParser",
-           "PandasReader", "execute_private_query"]
+
+__all__ = ["Rewriter", "QueryParser", "execute_private_query"]
